frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Nguyen Nhat Minh <52833936+Doki064@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^0.82.0"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
cchardet = "^2.1.7"
aiodns = "^3.0.0"
brotlipy = "^0.7.0"
orjson = "^3.5.3"
pandas = "^1.2.4"
pandas-profiling = "^3.0.0"
hiplot = "^0.1.25"
plotly = "^4.14.3"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
